+++ modules/config.py
@@ -11,7 +11,8 @@ from . import util
 
 # Settings dictionary to build website
 settings_dict = {
-    "version": "1.1.0",
+    "catalog_version": "6.2",
+    "website_version": "1.1",
     "banner_enabled": "false",
     "banner_message": "<a href='https://www.mitre.org/attackcon-streamed-live' target='_blank'> Register to stream ATT&CKcon 2.0 October 29-30</a>",
     "domains": ["pre-attack", "enterprise-attack", "mobile-attack"],