+++ update-attack.py
@@ -245,7 +245,8 @@ def generate_base_template():
                              "{% set BANNER_MESSAGE = \"${banner_message}\" %}\n"
                              "{% set NAVIGATION_MENU = ${navigation} -%}\n"
                              "{% set DOMAINS = ${domains} -%}\n"
-                             "{% set VERSION = \"v${version}\" -%}\n"
+                             "{% set CATALOG_VERSION = \"${catalog_version}\" -%}\n"
+                             "{% set WEBSITE_VERSION = \"${website_version}\" -%}\n"
                              "{% set active_page = active_page|"
                              "default('index') -%}\n")
     
@@ -262,7 +263,8 @@ def generate_base_template():
                                                             replace("\"", "'")
         base_dict['navigation'] = config.settings_dict['navigation_menu']
         base_dict['domains'] = config.settings_dict['domain_aliases']
-        base_dict['version'] = config.settings_dict['version']
+        base_dict['catalog_version'] = config.settings_dict['catalog_version']
+        base_dict['website_version'] = config.settings_dict['website_version']
         jinja_settings = base_template.substitute(base_dict)
   
     with open(base_template_path, 'w+') as f: